Adds a unit test.

+++ locationservice/naive.go
@@ -17,13 +17,13 @@ func (n *NaiveLocationService) Add(id string, point GeoPoint) {
 	n.locations = append(n.locations, pointContainer{id, toS2Point(point)})
 }
 
-func (n *NaiveLocationService) FindClosest(point GeoPoint, maxDistanceMeters float64) []SearchResult {
+func (n *NaiveLocationService) FindClosest(point GeoPoint, maxDistanceMeters int) []SearchResult {
 	target := toS2Point(point)
 
 	result := []SearchResult{}
 	for _, location := range n.locations {
 		distance := target.Distance(location.point).Radians() * earthRadiusMeters
-		if distance <= maxDistanceMeters {
+		if int(distance) <= maxDistanceMeters {
 			result = append(result, SearchResult{location.id, int(distance)})
 		}
 	}
